Null Results
Uploading new null results guide.

The latest version is completely different from the first version. It's written by Jenny and previously reviewed by both NI and GB. This is the first time that this version is uploaded on Github.

Action request: Please make a decision whether this guide is ready for a final review.